refactor: move view definition to views.py
