+++ src/apps/organizations/admin.py
@@ -1,8 +1,6 @@
 from django.contrib import admin
 from django.urls import path
-from django.views.generic import TemplateView
 from unfold.contrib.filters.admin import ChoicesDropdownFilter
-from unfold.views import UnfoldModelAdminViewMixin
 
 from apps.methods.models import Method
 from apps.users.models import UserProfile
@@ -12,27 +10,7 @@ from project.decorators import gov_admin_register, register_with_default_templat
 from .forms import OrganizationAdminForm
 from .helpers import get_organization_method_filter
 from .models import Organization, Project
-
-
-# Custom view
-class RegistrationRequestView(UnfoldModelAdminViewMixin, TemplateView):
-    title = "Registration requests"
-    permission_required = ()
-    template_name = "admin/organizations/registration_requests.html"
-
-    def get_context_data(self, *args, **kwargs):
-        context = super().get_context_data(**kwargs)
-        if "q" in self.request.GET:
-            query_filter = self.request.GET["q"]
-            context["organizations"] = Organization.objects.filter(
-                name__contains=query_filter
-            ).exclude(status=Organization.Status.ACCEPTED)
-            context["query_filter"] = self.request.GET["q"]
-        else:
-            context["organizations"] = Organization.objects.all().exclude(
-                status=Organization.Status.ACCEPTED
-            )
-        return context
+from .views import RegistrationRequestView
 
 
 # Add superadmin views with default Unfold templates
@@ -123,13 +101,12 @@ class OrganizationAdmin(ModelAdmin):
 
     # Add custom urls
     def get_urls(self):
-        registration_requests_view = self.admin_site.admin_view(
-            RegistrationRequestView.as_view(model_admin=self)
-        )
         urls = super().get_urls() + [
             path(
                 "registration-requests",
-                registration_requests_view,
+                self.admin_site.admin_view(
+                    RegistrationRequestView.as_view(model_admin=self)
+                ),
                 name="registration_requests",
             ),
         ]
